Refactor UI components, optimize imports, and improve code structure
